rename rule name
