+++ sonar-plugin/javascript-checks/src/main/java/org/sonar/javascript/checks/CheckList.java
@@ -258,6 +258,7 @@ public final class CheckList {
       NoMimeSniffCheck.class,
       NoMisleadingArrayReverseCheck.class,
       NoMisusedNewCheck.class,
+      NoMisusedPromisesCheck.class,
       NoMixedContentCheck.class,
       NoNestedSwitchCheck.class,
       NoNestedTemplateLiteralsCheck.class,
@@ -274,7 +275,6 @@ public final class CheckList {
       NoSameArgumentAssertCheck.class,
       NoSparseArraysCheck.class,
       NoThisAliasCheck.class,
-      NoUnexpectedPromiseCheck.class,
       NoUniqKeyCheck.class,
       NoUnnecessaryTypeAssertionCheck.class,
       NoUnnecessaryTypeConstraintCheck.class,
