+++ src/linting/eslint/rules/index.ts
@@ -148,6 +148,7 @@ import { rule as noIpForward } from './no-ip-forward';
 import { rule as noLabels } from './no-labels';
 import { rule as noMimeSniff } from './no-mime-sniff';
 import { rule as noMisleadingArrayReverse } from './no-misleading-array-reverse';
+import { rule as noMisusedPromises } from './no-misused-promises';
 import { rule as noMixedContent } from './no-mixed-content';
 import { rule as noNestedAssignment } from './no-nested-assignment';
 import { rule as noNestedConditional } from './no-nested-conditional';
@@ -169,7 +170,6 @@ import { rule as noTryPromise } from './no-try-promise';
 import { rule as noUndefinedArgument } from './no-undefined-argument';
 import { rule as noUndefinedAssignment } from './no-undefined-assignment';
 import { rule as noUnenclosedMultilineBlock } from './no-unenclosed-multiline-block';
-import { rule as noUnexpectedPromise } from './no-unexpected-promise';
 import { rule as noUniqKey } from './no-uniq-key';
 import { rule as noUnsafeUnzip } from './no-unsafe-unzip';
 import { rule as noUnthrownError } from './no-unthrown-error';
@@ -384,6 +384,7 @@ rules['no-ip-forward'] = noIpForward;
 rules['no-labels'] = noLabels;
 rules['no-mime-sniff'] = noMimeSniff;
 rules['no-misleading-array-reverse'] = noMisleadingArrayReverse;
+rules['no-misused-promises'] = noMisusedPromises;
 rules['no-mixed-content'] = noMixedContent;
 rules['no-nested-assignment'] = noNestedAssignment;
 rules['no-nested-conditional'] = noNestedConditional;
@@ -405,7 +406,6 @@ rules['no-try-promise'] = noTryPromise;
 rules['no-undefined-argument'] = noUndefinedArgument;
 rules['no-undefined-assignment'] = noUndefinedAssignment;
 rules['no-unenclosed-multiline-block'] = noUnenclosedMultilineBlock;
-rules['no-unexpected-promise'] = noUnexpectedPromise;
 rules['no-uniq-key'] = noUniqKey;
 rules['no-unsafe-unzip'] = noUnsafeUnzip;
 rules['no-unthrown-error'] = noUnthrownError;
