+++ tests/linting/eslint/rules/comment-based/no-misused-promises.ts
